@@ -12,30 +12,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.8",
-    "@angular/common": "^5.2.8",
-    "@angular/compiler": "^5.2.8",
-    "@angular/core": "^5.2.8",
-    "@angular/forms": "^5.2.8",
-    "@angular/http": "^5.2.8",
-    "@angular/platform-browser": "^5.2.8",
-    "@angular/platform-browser-dynamic": "^5.2.8",
-    "@angular/router": "^5.2.8",
-    "angular2-jwt": "^0.2.3",
+    "@angular/animations": "6.1.3",
+    "@angular/common": "6.1.3",
+    "@angular/compiler": "6.1.3",
+    "@angular/core": "6.1.3",
+    "@angular/forms": "6.1.3",
+    "@angular/http": "6.1.3",
+    "@angular/platform-browser": "6.1.3",
+    "@angular/platform-browser-dynamic": "6.1.3",
+    "@angular/router": "6.1.3",
+    "@auth0/angular-jwt": "^2.0.0",
     "chart.js": "^2.7.2",
     "core-js": "^2.4.1",
     "font-awesome": "^4.7.0",
     "moment": "^2.18.1",
     "ng2-currency-mask": "^4.3.1",
     "ng2-toasty": "^4.0.3",
-    "primeng": "^5.2.3",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "primeicons": "^1.0.0-beta.10",
+    "primeng": "^6.1.2",
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.2.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "^5.2.8",
-    "@angular/language-service": "^5.2.8",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "6.1.4",
+    "@angular/compiler-cli": "6.1.3",
+    "@angular/language-service": "6.1.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -51,7 +54,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.4.2"
+    "typescript": "2.9.2"
   },
   "engines": {
     "node": "8.6.0",
