@@ -1,7 +1,7 @@
+import { MoneyHttp } from './../seguranca/money-http';
 import { Injectable } from '@angular/core';
 
 import { environment } from './../../environments/environment';
-import { AuthHttp } from 'angular2-jwt';
 import 'rxjs/add/operator/toPromise';
 
 @Injectable()
@@ -10,17 +10,14 @@ export class CategoriaService {
    categoriasUrl: string;
 
    constructor(
-      private http: AuthHttp,
+      private http: MoneyHttp,
    ) {
       this.categoriasUrl = `${environment.apiUrl}/categorias`;
     }
 
    listarTodas(): Promise<any> {
       return this.http.get(this.categoriasUrl)
-      .toPromise()
-      .then(resp => {
-         return resp.json();
-      });
+      .toPromise();
    }
 
 }
