@@ -1,7 +1,7 @@
+import { MoneyHttp } from './../seguranca/money-http';
 import { Injectable } from '@angular/core';
-import { ResponseContentType, URLSearchParams } from '@angular/http';
+import { HttpParams } from '@angular/common/http';
 
-import { AuthHttp } from 'angular2-jwt';
 import { environment } from './../../environments/environment';
 
 import * as moment from 'moment';
@@ -12,18 +12,18 @@ export class RelatoriosService {
 
    lancamentosUrl: string;
 
-  constructor(private http: AuthHttp) {
+  constructor(private http: MoneyHttp) {
      this.lancamentosUrl = `${environment.apiUrl}/lancamentos`;
   }
 
    relatorioLancamentosPorPessoa(inicio: Date, fim: Date) {
-      const params = new URLSearchParams();
-      params.set('inicio', moment(inicio).format('YYYY-MM-DD'));
-      params.set('fim', moment(fim).format('YYYY-MM-DD'));
+      const params = new HttpParams()
+         .append('inicio', moment(inicio).format('YYYY-MM-DD'))
+         .append('fim', moment(fim).format('YYYY-MM-DD'));
+
       return this.http.get(`${this.lancamentosUrl}/relatorios/por-pessoa`,
-      { search: params, responseType: ResponseContentType.Blob})
-         .toPromise()
-         .then(response => response.blob());
+      { params, responseType: 'blob'})
+         .toPromise();
    }
 
 }
