@@ -1,5 +1,5 @@
-import { JwtHelper } from 'angular2-jwt';
-import { Http, Headers } from '@angular/http';
+import { JwtHelperService } from '@auth0/angular-jwt';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 
 import { environment } from './../../environments/environment';
@@ -12,29 +12,28 @@ export class AuthService {
    jwtPayload: any;
 
    constructor(
-      private http: Http,
-      private jwtHelper: JwtHelper
+      private http: HttpClient,
+      private jwtHelper: JwtHelperService
    ) {
       this.oauthTokenUrl = `${environment.apiUrl}/oauth/token`;
       this.carregarToken();
     }
 
    login(usuario: string, senha: string): Promise<void> {
-      const headers = new Headers();
-      headers.append('Content-Type', 'application/x-www-form-urlencoded');
-      headers.append('Authorization', 'Basic YW5ndWxhcjpAbmd1bEByMA==');
+      const headers = new HttpHeaders()
+            .append('Content-Type', 'application/x-www-form-urlencoded')
+            .append('Authorization', 'Basic YW5ndWxhcjpAbmd1bEByMA==');
 
       const body = `username=${usuario}&password=${senha}&grant_type=password&client=angular`;
-      return this.http.post(this.oauthTokenUrl, body,
+      return this.http.post<any>(this.oauthTokenUrl, body,
                { headers, withCredentials: true })
          .toPromise()
          .then(response => {
-            this.armazenarToken(response.json().access_token);
+            this.armazenarToken(response.access_token);
          })
          .catch(error => {
             if (error.status === 400) {
-               const errorJson = error.json();
-               if (errorJson.error === 'invalid_grant') {
+               if (error.error === 'invalid_grant') {
                   return Promise.reject('Usuário ou senha inválida!');
                }
             }
@@ -43,17 +42,17 @@ export class AuthService {
    }
 
    obterNovoAccessToken(): Promise<void> {
-      const headers = new Headers();
-      headers.append('Content-Type', 'application/x-www-form-urlencoded');
-      headers.append('Authorization', 'Basic YW5ndWxhcjpAbmd1bEByMA==');
+      const headers = new HttpHeaders()
+            .append('Content-Type', 'application/x-www-form-urlencoded')
+            .append('Authorization', 'Basic YW5ndWxhcjpAbmd1bEByMA==');
 
       const body = 'grant_type=refresh_token';
 
-      return this.http.post(this.oauthTokenUrl, body,
+      return this.http.post<any>(this.oauthTokenUrl, body,
                { headers, withCredentials: true })
          .toPromise()
          .then(response => {
-            this.armazenarToken(response.json().access_token);
+            this.armazenarToken(response.access_token);
             return Promise.resolve(null);
          })
          .catch(error => {
