@@ -1,4 +1,4 @@
-import { AuthHttp } from 'angular2-jwt';
+import { MoneyHttp } from './money-http';
 import { Injectable } from '@angular/core';
 import { AuthService } from './auth.service';
 import { environment } from './../../environments/environment';
@@ -9,7 +9,7 @@ export class LogoutService {
    tokensRevokeUrl: string;
 
   constructor(
-     private http: AuthHttp,
+     private http: MoneyHttp,
      private auth: AuthService
    ) {
       this.tokensRevokeUrl = `${environment.apiUrl}/tokens/revoke`;
