@@ -4,26 +4,19 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Http, RequestOptions } from '@angular/http';
 
-import { AuthHttp, AuthConfig } from 'angular2-jwt';
+import { JwtModule } from '@auth0/angular-jwt';
 
 import { InputTextModule } from 'primeng/inputtext';
 import { ButtonModule } from 'primeng/button';
 
-import { MoneyHttp } from './money-http';
-import { AuthService } from './auth.service';
-import { SharedModule } from './../shared/shared.module';
 import { LoginFormComponent } from './login-form/login-form.component';
 import { SegurancaRoutingModule } from './seguranca-routing.module';
 import { AuthGuard } from './auth.guard';
 import { LogoutService } from './logout.service';
+import { environment } from '../../environments/environment';
 
-export function authHttpServiceFactory(auth: AuthService, http: Http, options: RequestOptions) {
-   const config = new AuthConfig({
-      globalHeaders: [
-         { 'Content-Type' : 'application/json' }
-      ]
-   });
-   return new MoneyHttp(auth, config, http, options);
+export function tokenGetter() {
+   return localStorage.getItem('token');
 }
 
 @NgModule({
@@ -32,6 +25,13 @@ export function authHttpServiceFactory(auth: AuthService, http: Http, options: R
     RouterModule,
     FormsModule,
 
+    JwtModule.forRoot({
+       config: {
+         tokenGetter: tokenGetter,
+         whitelistedDomains: environment.tokenWhitelistedDomains,
+         blacklistedRoutes: environment.tokenBlacklistedRoutes
+       }
+    }),
     InputTextModule,
     ButtonModule,
 
@@ -41,11 +41,6 @@ export function authHttpServiceFactory(auth: AuthService, http: Http, options: R
      LoginFormComponent
   ],
   providers: [
-     {
-         provide: AuthHttp,
-         useFactory: authHttpServiceFactory,
-         deps: [AuthService, Http, RequestOptions]
-     },
      AuthGuard,
      LogoutService
   ]
